Issue #7133:remove integer casting for invchead_invcnumber
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@2964 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/printLabelsByInvoice.cpp
@@ -151,7 +151,7 @@ void printLabelsByInvoice::sParseInvoiceNumber()
   query.prepare( "SELECT invchead_id "
                  "FROM invchead "
                  "WHERE (invchead_invcnumber=:invoiceNumber);" );
-  query.bindValue(":invoiceNumber", _invoiceNumber->text().toInt());
+  query.bindValue(":invoiceNumber", _invoiceNumber->text());
   query.exec();
   if (query.first())
   {
